move type files to folder
