--- types/global.d.ts
@@ -1,3 +1,0 @@
-// Global definitions for developement
-
-declare module '@volusion/element-proptypes';
